Generate HTML Tree for JSON Schema

This function takes a JSON schema file as input and generates an interactive HTML tree representation of the schema using a force-directed graph.

Usage

plot_json_schema(schema_file)

Arguments

schema_file

A string representing the path to the JSON schema file.

Value

An HTML widget containing the interactive graph.

Examples

schema_file <- system.file("schemas", "user_schema.json", package = "PIFDataModel")
plot_json_schema(schema_file)